implement showing the comments left about a trading equipment
