+++ practice-app/frontend/mercatus/src/App.js
@@ -5,6 +5,7 @@ import TradingEquipment from './TradingEquipment';
 import axios from 'axios';
 
 axios.defaults.baseURL = 'http://127.0.0.1:5000';
+axios.defaults.headers.post['Content-Type'] = 'text/json';
 
 function App() {
     return (
